Finished: reduction F. Next: create embedding for the extensions that have support greater than or equal to minsup

+++ PSM/calcSupport.h
@@ -4,5 +4,6 @@
 #include <cuda_runtime.h>
 #include <device_launch_parameters.h>
 #include "ExtensionStructure.h"
+#include "reduction.h"
 
-extern "C" inline cudaError_t calcSupport(Extension *d_UniqueExtension,unsigned int noElem_d_UniqueExtension,Extension *d_ValidExtension,unsigned int noElem_d_ValidExtension,int *d_scanB_Result,int *d_F,unsigned int noElem_F);
+extern "C" inline cudaError_t calcSupport(Extension *d_UniqueExtension,unsigned int noElem_d_UniqueExtension,Extension *d_ValidExtension,unsigned int noElem_d_ValidExtension,int *d_scanB_Result,float *d_F,unsigned int noElem_F);
